moved working gui code into main folder and removed matlab test folder
