--- src/matlab-test-code/ErrorLogTest.txt
@@ -1,15 +1,0 @@
-Initializing......
-
-Error logs currently reading input from file making sure it is running fine
-
-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-Reminder: make sure to remove the test button once error log is linked to the other windows in the GUI
-
-555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555555
-(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)(disc).(hello)
->>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-())))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))))
-
-????????????????????????????????????????????????????????????????????????????????????????????????
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
