--- src/matlab-test-code/Notes/1. Hello plot! — gtk3-matplotlib-cookbook 0.1 documentation.webloc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>URL</key>
-	<string>http://gtk3-matplotlib-cookbook.readthedocs.io/en/latest/hello-plot.html</string>
-</dict>
-</plist>
