--- src/matlab-test-code/Notes/user_interfaces example code- embedding_in_gtk3.py — Matplotlib .webloc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>URL</key>
-	<string>https://matplotlib.org/examples/user_interfaces/embedding_in_gtk3.html</string>
-</dict>
-</plist>
