--- src/matlab-test-code/Styling.css
@@ -1,153 +1,0 @@
-#blackBox {
-  background-color: black;
-}
-#redBox {
-  background-color: red;
-}
-.goldBox {
-  background-color: gold;
-}
-
-#blackBox {
-  background-color: black;
-}
-
-#ArmGuiLayoutWindow{
-  background-color: #403e3e;
-}
-
-#Yaw-left, #Yaw-Right{
-    color: white;
-    border-color: green;
-    background-color: green;
-    background-image:none;
-}
-
-#Yaw-Right{
-    border-radius: 0 52px 8px 0;
-    padding-left: 0px;
-    margin-left:0px
-}
-
-#Yaw-left{
-    border-radius: 52px 0 0 8px;
-    padding-right: 0px;
-    margin-right:0px
-    }
-
-#hposbtn{
-border-radius: 30px 30px 30px 30px
-}
-
-#stepper-motor-1-angle{
-    color: red;
-}
-
-#Claw-Open, #Claw-Close{
-    color: white;
-    border-color: #fcb315;
-    background-color: #fcb315;
-    background-image:none;
-}
-
-#Pitch-Down, #Pitch-Up{
-    color: white;
-    border-color: #30016f;
-    background-color: #30016f;
-    background-image:none;
-}
-
-
-#Claw-Controls{
-    color: white;
-    border-color: #71797A;
-    background-color: #71797A;
-    background-image:none;
-    border: 4px solid yellow;
-}
-
-#arm-controls{
-    color: white;
-    border-color: #254A4F;
-    background-color: #254A4F;
-    background-image:none;
-    border: 4px solid cyan;
-}
-
-#claw-control-title, #arm-controls-title{
-    font-weight: bold;
-}
-
-#Arm-Position-Label, #lidar, #error-logs-window, #motor-table-label,#Manual-Control{
-    font-weight: bold;color:white
-}
-
-#Arm-Down, #Arm-Up, #Arm-Left, #Arm-Right{
-    color: white;
-    border-color: #cc33ff;
-    background-color: #cc33ff;
-    background-image:none;
-}
-
-#Arm-Up-Hide, #Arm-Down-Hide{
-    opacity:0.2
-}
-
-#Arm-Back, #Arm-Forward{
-    color: white;
-    border-color: #ff0000;
-    background-color: #ff0000;
-    background-image:none;
-}
-
-#Open-Logs{
-	color: white;
-    border-color: green;
-    background-color: green;
-    background-image:none;
-}
-
-#Clear-Logs{
-	color: white;
-    border-color: red;
-    background-color: red;
-    background-image:none;
-}
-
-#motor-grid{
-	color: white;
-	border-color: #0000FF;
-	background-color: #2525a7;
-	background-image:none;
-	border: 4px solid #8A2BE2;padding: 52px 0 26px 0
-}
-
-#motor-status-label{
-	background-color: #8080ff;
-	font-weight: bold;
-}
-
-#motor-position-label{
-	background-color: #8080ff;
-	font-weight: bold;
-}
-
-#motor-current-label{
-	background-color: #8080ff;
-	font-weight: bold;
-}
-
-#stepper-motor-1-label,#stepper-motor-3-label, #arm-servo-motor-1-label{
-	background-color: #8080ff;
-}
-
-#stepper-motor-2-label, #stepper-motor-4-label, #arm-servo-motor-2-label
-{background-color: #4343AE;}
-
-#stepper-motor-1-label,#stepper-motor-2-label,#stepper-motor-3-label, #stepper-motor-4-label,
-#arm-servo-motor-1-label, #arm-servo-motor-2-label
-{font-weight: bold;}
-
-
-
-
